+++ app/models/usuario.py
@@ -1,4 +1,5 @@
-from sqlalchemy import *
+from sqlalchemy import Column, Integer, String, Boolean, DateTime
+
 from app import db
 
 
@@ -31,14 +32,3 @@ class Usuario(db.Model):
 
     def is_anonymous(self):
         return False
-
-
-class Participante(db.Model):
-    __tablename__ = 'participante'
-    id = Column(Integer, db.ForeignKey('usuario.id'), primary_key=True)
-    edicao = Column(Integer, nullable=False)
-    pacote = Column(Boolean, nullable=False)
-    pagamento = Column(Boolean, nullable=False)
-    camiseta = Column(String(3))
-    data_inscricao = Column(DateTime, nullable=False)
-    credenciado = Column(Boolean, nullable=False)
